Platform_Qualifier/requirements.txt
streamlit
pandas
openpyxl
requests
pyyaml
sentence-transformers